requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.6.15
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
Django==5.2.3
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
msgpack==1.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0